backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.0.350
openai==1.3.8
faiss-cpu==1.7.4
tavily-python==0.3.0
tenacity==8.2.3
pydantic==2.5.0
